e2e: permissions-scan-qr group
